address math support review comments
