+++ packages/react/src/components/code-block-view.module.css
@@ -14,8 +14,7 @@
   display: none;
 }
 
-/*REVIEW: do not call this class name as MathPreview. Currently we only have math using this, but in the future we might want to use mermaid too */
-.MathPreview {
+.Preview {
   display: block;
   padding: 0.25rem 0;
   overflow-x: auto;
