+++ packages/react/src/extensions/code-block-view.ts
@@ -1,5 +1,5 @@
+import { defineCodeBlockPreviewPlugin } from '@meowdown/core'
 import { union, type Extension } from '@prosekit/core'
-import { defineCodeBlockPreviewPlugin } from '@prosekit/extensions/code-block'
 import { defineReactNodeView, type ReactNodeViewComponent } from '@prosekit/react'
 
 import { CodeBlockView } from '../components/code-block-view.tsx'
@@ -12,7 +12,7 @@ export function defineCodeBlockView(): Extension {
       component: CodeBlockView satisfies ReactNodeViewComponent,
     }),
     // Decorates the code block under the caret; `CodeBlockView` reads it to
-    // decide whether the math preview shows alone or below the source.
-    defineCodeBlockPreviewPlugin(), // REVIEW: do not add @prosekit/extensions as a dependency of @meowdown/react just beccause you want to use `defineCodeBlockPreviewPlugin`. Re-export it from core.
+    // decide whether the preview shows alone or below the source.
+    defineCodeBlockPreviewPlugin(),
   )
 }
