--- packages/react/src/utils/math.ts
@@ -1,22 +1,0 @@
-import type { KaTeX } from '@meowdown/core'
-
-// REVIEW: DO NOT ADD this function `renderMathInto` to @meowdown/react. Just wrap it in core package and re-export. This logic is shared for both core and react.
-
-/**
- * Render TeX into `element` as native MathML (no KaTeX stylesheet or fonts
- * required). `throwOnError: false` renders parse errors as red text; the
- * catch covers the rare non-parse error so a bad formula can never crash a
- * render.
- */
-export function renderMathInto(
-  katex: KaTeX,
-  element: HTMLElement,
-  formula: string,
-  displayMode: boolean,
-): void {
-  try {
-    katex.render(formula, element, { displayMode, throwOnError: false, output: 'mathml' })
-  } catch (error) {
-    element.textContent = String(error)
-  }
-}
